Move Exceptions from metrics.util to util package
So I can reuse it for other new tests
