+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/metrics/RetryMetricTest.java
@@ -18,7 +18,7 @@
  */
 package org.eclipse.microprofile.fault.tolerance.tck.metrics;
 
-import static org.eclipse.microprofile.fault.tolerance.tck.metrics.util.Exceptions.expectTestException;
+import static org.eclipse.microprofile.fault.tolerance.tck.util.Exceptions.expectTestException;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
